refactor(token-registry): rename token registry status import to correct file path and create new types file
